standardizing asset compilation config
In staging and production, JS and CSS are minified. Sourcemaps are included on staging.

NOTE: Still seems like some Ember Data sourcemaps are included on production. Hopefully this will go away once we upgrade to the latest version of Ember Data, which is packaged as an EmberCLI add-on (rather than the current, weird way Ember Data is included in both `package.json` and `bower.json`.)